services/atlas-worker/pyproject.toml
[tool.poetry]
name = "atlas-worker"
version = "0.1.0"
description = "Atlas Worker for architecture generation"
authors = ["Infra Crew <infra@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.20"
fastapi = "^0.110"
google-cloud-pubsub = "^2.20"
httpx = "^0.27"
prometheus-client = "^0.20"
tenacity = "^8.2"
asyncio = "^3.4"

[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"